+++ src/app/components/forms/postEditForm/postEditForm.component.ts
@@ -33,13 +33,25 @@
 // Angular imports
 import { Component, Input, Output, EventEmitter, OnInit } from "@angular/core";
 import { FormBuilder, Validators } from "@angular/forms";
+import { map, mergeMap, of } from "rxjs";
 
 // App-related import
 import { Post } from "@app/interfaces/post.interface";
-import { ItemsService } from "@app/services/items.service";
 import { AdminService } from "@app/services/admin.service";
 import { ValidationService } from "@app/services/validation.service";
 import { AlertsService } from "@app/services/alerts.service";
+import { ApiClientService } from "@app/services/apiClient.service";
+
+interface PostEditResponse {
+  success: boolean;
+  updated: Post;
+}
+
+interface PostAndReportResponse {
+  success: boolean;
+  postId?: number;
+  reportId?: number;
+}
 
 @Component({
   selector: "post-edit-form",
@@ -58,10 +70,10 @@ export class PostEditForm implements OnInit {
 
   // CTOR
   constructor(
-    private itemsService: ItemsService,
     private adminService: AdminService,
     private validationService: ValidationService,
     private alertService: AlertsService,
+    private apiClient: ApiClientService,
     private fb: FormBuilder,
   ) {}
 
@@ -82,7 +94,6 @@ export class PostEditForm implements OnInit {
   editPost(e: Event, closeReport: boolean | null) {
     e.preventDefault();
 
-    const serviceToUse = closeReport === null ? "itemsService" : "adminService";
     const newText = this.postEditForm.controls.postText.value || "";
 
     // if the post is invalid, show an error message
@@ -94,28 +105,53 @@ export class PostEditForm implements OnInit {
       return;
     }
 
-    // if the post is valid, edit the text
-    // if there isn't a value for closeReport, it means it's sent from the regular edit
-    if (closeReport === null) {
-      this.editedItem.text = newText;
-      this.itemsService.editPost(this.editedItem);
-      // otherwise if there's a value it's coming from admin dashboard editing
+    this.editedItem.text = newText;
+
+    // Edit the post
+    this.apiClient
+      .patch<PostEditResponse>(`posts/${this.editedItem.id}`, this.editedItem)
+      .pipe(mergeMap((postResponse) => this.updateReportIfNecessary(closeReport, postResponse)))
+      .subscribe({
+        next: (response: PostAndReportResponse) => {
+          this.editMode.emit(false);
+          const editMessage = response.reportId
+            ? `Report ${response.reportId} was closed, and the associated post was edited!`
+            : `Post ${response.postId} was edited.`;
+
+          this.alertService.createSuccessAlert(`${editMessage} Refresh to view the updated post.`, {
+            reload: closeReport || true,
+          });
+        },
+      });
+  }
+
+  /**
+   * Updates the report if there's a report to update and the user chooses to.
+   * @param closeReport - whether to close the report.
+   * @param postResponse - the post edit response from the API.
+   * @returns
+   */
+  updateReportIfNecessary(closeReport: boolean | null, postResponse: PostEditResponse) {
+    // If there's a Close Report value and the admin selected
+    // to close it, also close the report.
+    if (closeReport === true) {
+      return this.adminService
+        .closeReport(this.reportData.reportID, false, postResponse.updated.id)
+        .pipe(
+          map((reportResponse) => {
+            return {
+              success: reportResponse.success,
+              postId: postResponse.updated.id,
+              reportId: reportResponse.updated.id,
+            };
+          }),
+        );
     } else {
-      let post = {
-        text: newText,
-        id: this.reportData.postID,
-      };
-
-      this.adminService.editPost(post, closeReport, this.reportData.reportID);
+      return of({
+        success: postResponse.success,
+        postId: postResponse.updated.id,
+        reportId: undefined,
+      });
     }
-
-    // check whether the post's data was updated in the database
-    this[serviceToUse].isUpdated.subscribe((value: Boolean) => {
-      // if it has, close the popup; otherwise, leave it on so that the user
-      // can fix whatever errors they have and try again
-      if (value) {
-        this.editMode.emit(false);
-      }
-    });
   }
 }